modularize polling logic
